Exact cover (#19)
* abstract impl of qaoa, exactcover impl.

* added max_k_cut

* fixed bug

* Add route weights.

* Add classical exactcover solver

* Add large instance.

* added success prob and exact cover 14 instance
